refactor(decomp): hoist raptor/config + constants into proximadb-raptor-common
Move `raptor/config.rs` (296 LOC) + `raptor/constants.rs` (252 LOC) into the
`proximadb-raptor-common` crate as `config` + `constants` modules (alongside the
`common` types hoisted by #955). Root `config.rs` + `constants.rs` become one-line
re-export shims, so every `crate::storage::engines::raptor::{config,constants}::*`
path resolves unchanged.

Both are clean shared types (no `crate::` deps — `config` uses only
`proximadb-compression-types`; `constants` is pure consts) and widely used
(`config`: 6 raptor modules; `constants`: 4), so the fold is mechanical and the
reroute is shim-only.

Prerequisite for the `consolidated_reader` extraction (it uses `super::config` +
`super::constants`, which must be crate-accessible for the move). ~548 LOC further
into the raptor-common crate; storage up-edge ratchet 181 (held — no crate:: deps
to remove).

Verified: root --lib + raptor-common + clippy --lib --bins clean; fmt clean;
layering + workspace-boundaries pass; work-commit-check green.
