+++ crates/storage/proximadb-raptor-common/src/config.rs
@@ -1,0 +1,296 @@
+use super::constants;
+use serde::{Deserialize, Serialize};
+
+// Foundation compression types
+use proximadb_compression_types::CompressionAlgorithm as FoundationCompressionAlgorithm;
+
+/// Accuracy level for search operations
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub enum AccuracyLevel {
+    /// 99.5% recall target - maximum accuracy
+    Maximum,
+    /// 98% recall target - balanced accuracy/performance
+    Balanced,
+    /// 94% recall target - optimized for speed
+    Fast,
+}
+
+/// P×K storage strategy based on K/D relationship
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum PxKStrategy {
+    /// Full dense storage (k < √d)
+    DenseFull,
+    /// Full storage with compression (√d ≤ k < d/4)
+    DenseCompressed,
+    /// Sparse storage with coverage ratio
+    SparseCoverage {
+        coverage: f32,
+        compression: CompressionStrategy,
+    },
+    /// Learned index for very large K
+    LearnedIndex,
+}
+
+/// Compression strategy for P×K storage
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum CompressionStrategy {
+    /// No compression - f32 (4 bytes)
+    Uncompressed,
+    /// Half precision - f16 (2 bytes)
+    Float16,
+    /// 8-bit quantization (1 byte)
+    Quantized8,
+    /// 6-bit quantization (0.75 bytes)
+    Quantized6,
+    /// 4-bit quantization (0.5 bytes)
+    Quantized4,
+    /// Delta encoding for sorted distances
+    DeltaEncoded,
+    /// Bit packing based on range
+    BitPacked,
+}
+
+#[derive(Debug, Clone)]
+pub struct RaptorConfig {
+    // Storage settings
+    pub rowgroup_size: usize,
+    pub compression: RaptorCompressionConfig,
+    pub use_proximaencoder: bool, // Enable Proxima SIMD encoding
+
+    // SIMD settings
+    pub enable_simd: bool,
+    pub simd_lanes: usize,
+
+    // Cloud I/O settings
+    pub enable_range_reads: bool,
+    pub prefetch_size_mb: usize,
+    pub cache_size_mb: usize,
+    pub cache_eviction_policy: EvictionPolicy,
+
+    // IVF clustering settings for RAPTOR's p²+k×p algorithm
+    pub enable_clustering: bool,
+    pub num_clusters: Option<usize>, // k value, defaults to √n if not specified
+    #[allow(dead_code)]
+    pub target_rowgroup_size: Option<usize>, // p value, auto-calculated if not specified
+    pub use_component_boosting: bool, // Enable distance component boosting
+
+    // Metadata settings
+    pub enable_complex_types: bool,
+    pub enable_bloom_filters: bool,
+    pub bloom_fpp: f64,
+    pub enable_statistics: bool,
+
+    // Vector settings
+    pub dimension: usize, // Required dimension from collection config
+
+    // Compaction settings
+    pub compaction_threshold_files: usize,
+    pub compaction_min_size_mb: usize,
+    pub enable_clustering_aware_compaction: bool,
+    pub compaction_config: Option<RaptorCompactionConfig>,
+    pub clustering_config: Option<RaptorClusteringConfig>,
+
+    // Performance settings
+    pub max_parallel_reads: usize,
+    pub buffer_pool_size_mb: usize,
+    pub enable_prefetching: bool,
+}
+
+/// RAPTOR-specific compression configuration with level support
+///
+/// Wraps foundation compression types with RAPTOR-specific level parameters
+/// for fine-tuned compression control.
+#[derive(Debug, Clone)]
+pub enum RaptorCompressionConfig {
+    /// No compression
+    None,
+    /// LZ4 compression (fast, low compression)
+    Lz4,
+    /// Zstandard compression with level
+    Zstd(i32), // compression level (1-22)
+    /// Snappy compression (fast, moderate compression)
+    Snappy,
+    /// Gzip compression with level
+    Gzip(u32), // compression level (1-9)
+}
+
+impl From<RaptorCompressionConfig> for FoundationCompressionAlgorithm {
+    fn from(config: RaptorCompressionConfig) -> Self {
+        match config {
+            RaptorCompressionConfig::None => FoundationCompressionAlgorithm::None,
+            RaptorCompressionConfig::Lz4 => FoundationCompressionAlgorithm::Lz4,
+            RaptorCompressionConfig::Zstd(_) => FoundationCompressionAlgorithm::Zstd,
+            RaptorCompressionConfig::Snappy => FoundationCompressionAlgorithm::Snappy,
+            RaptorCompressionConfig::Gzip(_) => FoundationCompressionAlgorithm::Gzip,
+        }
+    }
+}
+
+/// Legacy type alias for backward compatibility
+/// TODO: Migrate all uses to RaptorCompressionConfig (Phase 3.1)
+pub type CompressionCodec = RaptorCompressionConfig;
+
+#[derive(Debug, Clone)]
+pub enum EvictionPolicy {
+    Lru,
+    Lfu,
+    Arc,
+    Cost, // Cost-aware eviction based on I/O cost
+}
+
+/// Backwards-compat alias for [`RaptorCompactionConfig`].
+pub type CompactionConfig = RaptorCompactionConfig;
+
+#[derive(Debug, Clone)]
+pub struct RaptorCompactionConfig {
+    pub max_level: usize,             // For RAPTOR: always 0 (single level)
+    pub l0_trigger_file_count: usize, // For RAPTOR: 2 (compact when > 1 file)
+    pub target_file_size: usize,      // For RAPTOR: usize::MAX (single file)
+}
+
+/// Backwards-compat alias for [`RaptorClusteringConfig`].
+pub type ClusteringConfig = RaptorClusteringConfig;
+
+#[derive(Debug, Clone)]
+pub struct RaptorClusteringConfig {
+    pub num_clusters: usize,          // k value in p²+k×p
+    pub rowgroup_size: usize,         // p value in p²+k×p
+    pub boosting_alpha_own: f32,      // α₁ weight for own centroid
+    pub boosting_alpha_inter: f32,    // α₂ weight for inter-centroid
+    pub boosting_alpha_variance: f32, // α₃ weight for variance
+}
+
+impl Default for RaptorConfig {
+    fn default() -> Self {
+        // Smart defaults optimized for HNSW + Columnar architecture
+        Self {
+            // RowGroup size optimized for typical k<10 queries:
+            // - 1000 vectors balances I/O efficiency vs wasted reads
+            // - At k=10, worst case reads 1000 vectors for 10 results (1% efficiency)
+            // - At k=100, may need 2-3 rowgroups (acceptable)
+            // - Memory: ~4MB per rowgroup @ 1024-dim (fits in L3 cache)
+            // - HNSW local graph: ~16K edges (1000 nodes * 16 connections)
+            // - Sweet spot: minimizes wasted I/O while maintaining locality
+            rowgroup_size: constants::clustering::DEFAULT_ROWGROUP_SIZE,
+
+            // Compression optimized for vector data:
+            // - Zstd level 3 gives 2-3x compression with fast decompression
+            // - Applied per-column for selective decompression
+            // - Graph edges use dictionary encoding
+            compression: RaptorCompressionConfig::Zstd(constants::compression::DEFAULT_ZSTD_LEVEL),
+            use_proximaencoder: true, // Enable Proxima for SIMD-optimized encoding
+
+            enable_simd: true,
+            simd_lanes: constants::memory::DEFAULT_SIMD_LANES,
+
+            enable_range_reads: true,
+            prefetch_size_mb: constants::memory::DEFAULT_PREFETCH_SIZE_MB,
+            cache_size_mb: constants::memory::DEFAULT_CACHE_SIZE_MB,
+            cache_eviction_policy: EvictionPolicy::Cost,
+
+            // IVF clustering for RAPTOR's p²+k×p algorithm
+            // Automatically calculates optimal k and p values based on dataset size
+            enable_clustering: true,
+            num_clusters: None,           // Auto-calculate as √n
+            target_rowgroup_size: None,   // Auto-calculate based on L3 cache
+            use_component_boosting: true, // Enable advanced distance components
+
+            enable_complex_types: true,
+            enable_bloom_filters: true,
+            bloom_fpp: constants::compression::DEFAULT_BLOOM_FPP,
+            enable_statistics: true,
+
+            // Vector settings
+            dimension: constants::dimensions::DEFAULT, // Default to common embedding dimension, will be overridden by collection config
+
+            // Aggressive compaction for HNSW graph consistency:
+            // - Trigger at 2 files to maintain single navigable graph
+            // - No size threshold - compact immediately
+            // - Single level (L0) to avoid graph fragmentation
+            // - Supports 100GB+ files through columnar streaming
+            compaction_threshold_files: constants::io::COMPACTION_THRESHOLD_FILES,
+            compaction_min_size_mb: constants::io::COMPACTION_MIN_SIZE_MB,
+            enable_clustering_aware_compaction: true,
+            compaction_config: Some(RaptorCompactionConfig {
+                max_level: constants::io::MAX_LSM_LEVEL,
+                l0_trigger_file_count: constants::io::COMPACTION_THRESHOLD_FILES,
+                target_file_size: constants::io::TARGET_FILE_SIZE,
+            }),
+            clustering_config: Some(RaptorClusteringConfig {
+                num_clusters: constants::clustering::DEFAULT_CLUSTER_COUNT,
+                rowgroup_size: constants::clustering::DEFAULT_ROWGROUP_SIZE,
+                boosting_alpha_own: constants::boosting::ALPHA_OWN_DEFAULT,
+                boosting_alpha_inter: constants::boosting::ALPHA_INTER_DEFAULT,
+                boosting_alpha_variance: constants::boosting::ALPHA_VARIANCE_DEFAULT,
+            }),
+
+            max_parallel_reads: constants::memory::DEFAULT_MAX_PARALLEL_READS,
+            buffer_pool_size_mb: constants::memory::DEFAULT_BUFFER_POOL_SIZE_MB,
+            enable_prefetching: true,
+        }
+    }
+}
+
+impl RaptorConfig {
+    /// Configuration optimized for small k queries (k<10)
+    pub fn for_small_k() -> Self {
+        Self {
+            rowgroup_size: 500,  // Minimize wasted reads
+            cache_size_mb: 4096, // Cache more rowgroups
+            ..Default::default()
+        }
+    }
+
+    /// Configuration optimized for medium k queries (k~100)
+    pub fn for_medium_k() -> Self {
+        Self {
+            rowgroup_size: 2000, // Balance I/O and efficiency
+            ..Default::default()
+        }
+    }
+
+    /// Configuration optimized for large k queries (k>100)
+    pub fn for_large_k() -> Self {
+        Self {
+            rowgroup_size: 5000, // Maximize throughput
+            enable_prefetching: true,
+            prefetch_size_mb: 128,
+            ..Default::default()
+        }
+    }
+
+    pub fn for_cloud() -> Self {
+        Self {
+            enable_range_reads: true,
+            prefetch_size_mb: 64,
+            cache_size_mb: 2048,
+            compression: RaptorCompressionConfig::Zstd(6),
+            rowgroup_size: 1000, // Default for cloud
+            ..Default::default()
+        }
+    }
+
+    pub fn for_local_ssd() -> Self {
+        Self {
+            enable_range_reads: false,
+            prefetch_size_mb: 16,
+            cache_size_mb: 512,
+            compression: RaptorCompressionConfig::Lz4,
+            max_parallel_reads: 16,
+            ..Default::default()
+        }
+    }
+
+    pub fn for_high_performance() -> Self {
+        Self {
+            rowgroup_size: 5000,
+            compression: RaptorCompressionConfig::None,
+            enable_simd: true,
+            simd_lanes: 32,
+            cache_size_mb: 4096,
+            max_parallel_reads: 32,
+            buffer_pool_size_mb: 2048,
+            ..Default::default()
+        }
+    }
+}
